% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseVectorFieldFunctions.R
\name{baseVectorFields}
\alias{baseVectorFields}
\title{Base Vector Fields}
\usage{
baseVectorFields(t, curPos)
}
\arguments{
\item{t}{a numeric scalar. The current time value.}

\item{curPos}{a d-dimensional numeric vector. The current position in the field.}
}
\value{
a d x m matrix. Each column corresponds to the vector for one model.
}
\description{
A base vector field function.
This is a template for creating any vector field function for using this package.
}
\examples{
baseVectorFields(0, c(1,1))
}
