% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructIntegralLists_Traj.R
\name{constructIntegralList_PathTraj}
\alias{constructIntegralList_PathTraj}
\title{constructIntegralList_PathTraj}
\usage{
constructIntegralList_PathTraj(
  EstPath,
  baseVectorFields,
  cur_integral_list,
  worst_split,
  Path_TimeSplits,
  Traj_TimeSplits,
  n_dim,
  n_models
)
}
\arguments{
\item{EstPath}{an EstimatedPath object.}

\item{baseVectorFields}{a baseVectorFields function.}

\item{cur_integral_list}{a list of matrices. Each corresponds with a block of the previous overall PathTrajectory integral matrix.}

\item{worst_split}{a numeric scalar. The region to be split.}

\item{Path_TimeSplits}{a numeric vector. The nodes of the path spline with the starting and ending times.}

\item{Traj_TimeSplits}{a numeric vector. The nodes of the currrent trajectory spline with the starting and ending times.}

\item{n_dim}{a numeric scalar. The number of dimensions in the path.}

\item{n_models}{a numeric scalar. The number of vector fields being mixed.}
}
\value{
a list of matrices.
}
\description{
constructIntegralList_PathTraj
}
