% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructIntegralLists_Traj.R
\name{constructIntegralList_Traj}
\alias{constructIntegralList_Traj}
\title{constructIntegralList_Traj}
\usage{
constructIntegralList_Traj(
  EstPath,
  baseVectorFields,
  cur_integral_list,
  time_splits,
  worst_split,
  n_models
)
}
\arguments{
\item{EstPath}{an EstimatedPath object.}

\item{baseVectorFields}{a baseVectorField function.}

\item{cur_integral_list}{a list of matrices. Each corresponds with a block of the previous overall trajectory integral matrix.}

\item{time_splits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{worst_split}{a numeric scalar. The region to be split.}

\item{n_models}{a numeric scalar. The number of vector fields being mixed.}
}
\value{
a list of matrices.
}
\description{
constructIntegralList_Traj
}
