% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructIntegralMatrixList_Path.R
\name{constructIntegralMatrixList_Path}
\alias{constructIntegralMatrixList_Path}
\title{constructIntegralMatrixList_Path}
\usage{
constructIntegralMatrixList_Path(cur_integral_list, time_splits, worst_split)
}
\arguments{
\item{cur_integral_list}{a list of matrices. Each corresponds with a block of the previous overall integral matrix.}

\item{time_splits}{a numeric vector. The nodes of the current spline with the starting and ending times.}

\item{worst_split}{a numeric scalar. The region to be split.}
}
\value{
a list of matrices.
}
\description{
constructIntegralMatrixList_Path
}
