% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundedPathRegressionFunctions.R
\name{constructUnboundedCubicSplineModelMatrix_Path}
\alias{constructUnboundedCubicSplineModelMatrix_Path}
\title{constructUnboundedCubicSplineModelMatrix_Path}
\usage{
constructUnboundedCubicSplineModelMatrix_Path(
  Data,
  time_splits,
  split_labels,
  V_Smooth = F
)
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{time_splits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{split_labels}{a numeric or factor vector. Labels for each point in Data for which region of the spline the point falls into.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}
}
\value{
a numeric matrix. The regression model matrix X
}
\description{
Constructing the model matrix for the path cubic spline regression.
}
