% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundedPathRegressionFunctions.R
\name{constructUnboundedCubicSplineTransitionMatrix_Path}
\alias{constructUnboundedCubicSplineTransitionMatrix_Path}
\title{constructUnboundedCubicSplineTransitionMatrix_Path}
\usage{
constructUnboundedCubicSplineTransitionMatrix_Path(time_splits, V_Smooth = T)
}
\arguments{
\item{time_splits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}
}
\value{
a numeric matrix.
}
\description{
Construct the matrix to transform the free parameters of the cubic spline into the full set of parameters.
}
