% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/constructUnboundedCubicSplineTransitionMatrix_Traj.R
\name{constructUnboundedCubicSplineTransitionMatrix_Traj}
\alias{constructUnboundedCubicSplineTransitionMatrix_Traj}
\title{constructUnboundedCubicSplineTransitionMatrix_Traj}
\usage{
constructUnboundedCubicSplineTransitionMatrix_Traj(
  time_splits,
  n_models,
  V_Smooth = T
)
}
\arguments{
\item{time_splits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{n_models}{a numeric scalar. The number of vector fields being mixed.}

\item{V_Smooth}{a boolean. TRUE if the trajectory should be estimated using C2 smoothness. FALSE if C1.}
}
\value{
a numeric matrix. Transforms the free trajectory coefficients into the full coefficients.
}
\description{
constructUnboundedCubicSplineTransitionMatrix_Traj
}
