% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrestritcedCubicRegression_Traj.R
\name{constructUnrestrictedCubicRegressionModelMatrix_Traj}
\alias{constructUnrestrictedCubicRegressionModelMatrix_Traj}
\title{constructUnrestrictedCubicRegressionModelMatrix_Traj}
\usage{
constructUnrestrictedCubicRegressionModelMatrix_Traj(
  cur_PathData,
  start_time,
  n_dim,
  n_models
)
}
\arguments{
\item{cur_PathData}{a numeric data frame. The path data to use to estimate the cubic trajectories.}

\item{start_time}{a numeric scalar. The centering time of the cubic function.}

\item{n_dim}{a numeric scalar. The number of dimensions in the path.}

\item{n_models}{a numeric scalar. The number of vector fields being mixed.}
}
\value{
a numeric matrix. The linear regression model matrix.
}
\description{
constructUnrestrictedCubicRegressionModelMatrix_Traj
}
