% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationFunctions.R
\name{coverageRateTraj}
\alias{coverageRateTraj}
\title{coverageRateTraj}
\usage{
coverageRateTraj(EstTraj, TrueTraj, TrueTrajTimeSplits, CL = 0.95)
}
\arguments{
\item{EstTraj}{an EstimatedTrajectory object. Must have been estimated using the Optimzation method.}

\item{TrueTraj}{a list of numeric matrices. The list of coefficient matrices of a true trajectory.}

\item{TrueTrajTimeSplits}{a numeric vector. The nodes of the true trajectory in addition with the starting and ending time.}

\item{CL}{a real number between 0 and 1. The confidence level of the intervals to check the coverage rate.}
}
\value{
a numeric vector. Each entry corresponds to the coverage rate for a given model
}
\description{
coverageRateTraj
}
