% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationFunctions.R
\name{cubicAvgSquareDist}
\alias{cubicAvgSquareDist}
\title{cubicAvgSquareDist}
\usage{
cubicAvgSquareDist(
  EstCubic,
  TrueCubic,
  EstTimeSplits,
  TrueTimeSplits,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{EstCubic}{a list of numeric matrices. The list of coefficient matrices of the estimated cubic spline.}

\item{TrueCubic}{a list of numeric matrices. The list of coefficient matrices of a true cubic spline.}

\item{EstTimeSplits}{a numeric vector. The nodes of the estimated cubic spline in addition with the starting and ending time.}

\item{TrueTimeSplits}{a numeric vector. The nodes of the true cubic spline in addition with the starting and ending time.}

\item{lower_bound}{a number scalar. The lower bound of the intergal.}

\item{upper_bound}{a number scalar. The upper bound of the intergal.}
}
\value{
a numeric scalar.
}
\description{
Finds the average squared distance between 2 cubic spline functions.
}
\examples{
cubicAvgSquareDist(EstCubic = ExEstTraj_Optimization$Traj, 
                   TrueCubic = ExTraj$Traj, 
                   EstTimeSplits = ExEstTraj_Optimization$TimeSplits, 
                   TrueTimeSplits = ExTraj$TimeSplits)
  
}
