% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicEvaluationFunctions.R
\name{evaluateCubicVelocity}
\alias{evaluateCubicVelocity}
\title{evaluateCubicVelocity}
\usage{
evaluateCubicVelocity(t, Cubic, startTime)
}
\arguments{
\item{t}{a numeric vector. The times at which the cubic functions should be evaluated.}

\item{Cubic}{an n x 4 numeric matrix.
The first columns contains the cubic coefficients (a), the second the quadratic (b), etc.}

\item{startTime}{a numeric scalar. The value of the centering time \eqn{t_0}.}
}
\value{
a numeric matrix. Each row corresponds to the outputs of each cubic function's first derivative at the time values.
}
\description{
Evaluate the first derivative of a series of centered cubic functions at a given time.
\deqn{f(t) = a(t - t_0)^3 + b(t - t_0)^2 + c(t - t_0) + d}
}
\examples{
evaluateCubicVelocity(1,c(1,1,1,1),0)
evaluateCubicVelocity(0:5, matrix(c(1,1,1,1,
                                    1,2,3,4,
                                    2,1,6,4), nrow = 3, byrow = TRUE), 10)
}
