% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExData}
\alias{ExData}
\title{Example Data}
\format{
\subsection{\code{ExData}}{

A matrix with 501 rows and 3 columns.
}
}
\source{
The example in the RungeKutta function with some error.
}
\usage{
ExData
}
\description{
An example data set
}
\keyword{datasets}
