% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExEstPath_ByFours}
\alias{ExEstPath_ByFours}
\title{Example Estimated Path}
\format{
\subsection{\code{ExEstPath_ByFours}}{

A list with 5 elements:
\describe{
\item{Path}{A list of cubic coefficient matrices.}
\item{TimeSplits}{A vector of node, starting, and ending times.}
\item{n_splits}{The number of regions in the cubic spline.}
\item{Error}{A matrix with the squared residuals in each dimension for each data point.}
\item{PosSplits}{A vector of boolean describing which regions in the spline can be split in the next iteration.}
}
}
}
\source{
The example in the getEstimatedPaths_ByFours function.
}
\usage{
ExEstPath_ByFours
}
\description{
A path estimated using the ByFours method.
}
\keyword{datasets}
