% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExEstPath_Projection}
\alias{ExEstPath_Projection}
\title{Example Estimated Path}
\format{
\subsection{\code{ExEstPath_Projection}}{

A list with 14 elements:
\describe{
\item{Path}{A list of cubic coefficient matrices.}
\item{TimeSplits}{A vector of node, starting, and ending times.}
\item{n_splits}{The number of regions in the cubic spline.}
\item{Error}{A matrix with the squared residuals in each dimension for each data point.}
\item{PosSplits}{A vector of boolean describing which regions in the spline can be split in the next iteration.}
\item{Cov}{A list of lists containing covariance matrices for each region and dimension of the cubic spline coefficients.}
\item{Sigma2}{A list of vectors containing the sigma2 estimates for each region and dimension of the cubic spline.}
\item{SplitLabels}{A factor vector describing which region of the spline each data point lies.}
\item{FullCov}{A list a covariance matrices for the cubic spline coefficients for each dimension.}
\item{Free_Parameters}{A matrix containing the estimates of the free coefficients for each dimension's spline.}
\item{Free_Parameter_Cov}{A list of covariance matrices for the free coefficients for each dimension's spline.}
\item{UnrestrictedCoef}{A list of unrestricted cubic coefficents without any induced smoothness in the spline.}
\item{UnrestrictedCov}{A list of lists containing covariance matrices for each region and dimension of the unrestricted cubic spline coefficients.}
\item{IntegralList}{A list of matrices used to project the unrestricted coefficients into the restricted space.}
}
}
}
\source{
The example in the getEstimatedPaths_Projection function.
}
\usage{
ExEstPath_Projection
}
\description{
A path estimated using the Projection method.
}
\keyword{datasets}
