% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExEstTraj_ByFours}
\alias{ExEstTraj_ByFours}
\title{Example Estimated Trajectory}
\format{
\subsection{\code{ExEstTraj_ByFours}}{

A list with 5 elements:
\describe{
\item{Traj}{A list of cubic coefficient matrices.}
\item{TimeSplits}{A vector of node, starting, and ending times.}
\item{n_splits}{The number of regions in the cubic spline.}
\item{Error}{A matrix with the squared residuals in each dimension for each data point.}
\item{PosSplits}{A vector of boolean describing which regions in the spline can be split in the next iteration.}
}
}
}
\source{
The example in the getEstimatedTrajectories_ByFours function.
}
\usage{
ExEstTraj_ByFours
}
\description{
A trajectory estimated using the ByFours method.
}
\keyword{datasets}
