% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExEstTraj_OneShot}
\alias{ExEstTraj_OneShot}
\title{Example Estimated Trajectory}
\format{
\subsection{\code{ExEstTraj_OneShot}}{

A list with 9 elements:
\describe{
\item{Traj}{A list of cubic coefficient matrices.}
\item{TimeSplits}{A vector of node, starting, and ending times.}
\item{n_splits}{The number of regions in the cubic spline.}
\item{Error}{A vector with the integral squared errors for each region of the spline.}
\item{PosSplits}{A vector of boolean describing which regions in the spline can be split in the next iteration.}
\item{Cov}{A list of lists containing covariance matrices for each region and model of the cubic spline coefficients.}
\item{Sigma2}{A list of vectors containing the sigma2 estimates for each region of the cubic spline.}
\item{IntegralList_Traj}{A list of matrices used to transform the path coefficients into the trajectory coefficients.}
\item{IntegralList_PathTraj}{A list of matrices used to transform the path coefficients into the trajectory coefficients.}
}
}
}
\source{
The example in the getEstimatedTrajectories_OneShot function.
}
\usage{
ExEstTraj_OneShot
}
\description{
A trajectory estimated using the OneShot method.
}
\keyword{datasets}
