% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExPath}
\alias{ExPath}
\title{Example Path}
\format{
\subsection{\code{ExPath}}{

A list with 2 elements:
\describe{
\item{Path}{A list of cubic coefficient matrices.}
\item{TimeSplits}{A vector of node, starting, and ending times.}
}
}
}
\source{
N/A
}
\usage{
ExPath
}
\description{
An example path
}
\keyword{datasets}
