% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExPropagationList}
\alias{ExPropagationList}
\title{Example PropagationList}
\format{
\subsection{\code{ExPropagationList}}{

A list with 2 elements:
\describe{
\item{First}{A data frame with the propagation estimates.}
\item{Second}{A list of matrices with the intermediate Runge Kutta steps for each propagation.}
}
}
}
\source{
The example in the getPropagationList function.
}
\usage{
ExPropagationList
}
\description{
An example propagation list used in the drifterVisualizationTool.
}
\keyword{datasets}
