% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExTraj}
\alias{ExTraj}
\title{Example Trajectory}
\format{
\subsection{\code{ExTraj}}{

A list with 2 elements:
\describe{
\item{Traj}{A list of cubic coefficient matrices.}
\item{TimeSplits}{A vector of node, starting, and ending times.}
}
}
}
\source{
N/A
}
\usage{
ExTraj
}
\description{
An example trajectory
}
\keyword{datasets}
