% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateC1Smoothness.R
\name{generateC1Smoothness}
\alias{generateC1Smoothness}
\title{Generate C1 Smoothness}
\usage{
generateC1Smoothness(
  prevCubic,
  curCubic,
  prevStartTime,
  curStartTime,
  meetTime
)
}
\arguments{
\item{prevCubic}{an nx4 numeric matrix.
Each row corresponds the coefficients of the cubic functions for which the current cubic will be connected.}

\item{curCubic}{an nx2 numeric matrix.
Each row corresponds the cubic and quadratic coefficients of the cubic functions currently being connected.}

\item{prevStartTime}{a numeric scalar. The centering time of the previous cubic functions.}

\item{curStartTime}{a numeric scalar. The centering time of the current cubic functions.}

\item{meetTime}{a numeric scalar. The time at which the previous and current cubic functions meet.}
}
\value{
an nx4 numeric matrix.
Each row corresponds the coefficients of the current cubic functions with C1 smoothness to the previous cubic functions.
}
\description{
Generate C1 Smoothness for a cubic function given a previous cubic to connect.
}
\examples{

prevCubic <- matrix(c(1,2,3,4), nrow = 1, byrow = TRUE)
curCubic <- matrix(c(2,1/2), nrow = 1, byrow = TRUE)
prevStartTime <- 0
curStartTime <- 1
meetTime <- 1

generateC1Smoothness(prevCubic, curCubic, prevStartTime, curStartTime, meetTime)

}
