% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTrajectoryFunctions.R
\name{generateFullTrajectory}
\alias{generateFullTrajectory}
\title{Generate Full Trajectory Function}
\usage{
generateFullTrajectory(TimeSplits, numModels, sigma, k, startWeights)
}
\arguments{
\item{TimeSplits}{a numeric vector. The spline node locations with the starting and ending times.}

\item{numModels}{a numeric scalar. The number of trajectories to generate.}

\item{sigma}{a non-negative numeric scalar. The standard deviation of the normal distribution generating the starting values of the trajectories. Not needed if startWeights is specified.}

\item{k}{a non-negative numeric scalar. The standard deviation of the induced normal distribution on difference between the starting and ending values of the individual cubic functions.
This is the equivalent of endSD in generateIndividualTrajectory function.}

\item{startWeights}{a numeric vector. The starting values of the trajectory functions.}
}
\value{
a list. Each element is a coefficient matrix corresponding to a segment of the spline.
}
\description{
A function to randomly generate full cubic spline trajectory functions.
}
\examples{

generateFullTrajectory(TimeSplits = 0:5, numModels = 2, sigma = 0.1, k = 0.5)

}
