% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRandomPath.R
\name{generateRandomPath}
\alias{generateRandomPath}
\title{generateRandomPath}
\usage{
generateRandomPath(free_param_mean, free_param_cov, TimeSplits, V_Smooth)
}
\arguments{
\item{free_param_mean}{a numeric matrix.
Each column corresponds to the means of the free parameters for a dimension's estimated path cubic spline.}

\item{free_param_cov}{a list of numeric matrices.
Each element corresponds to the covariance matrix of the free parameters for a dimension's estimated path cubic spline}

\item{TimeSplits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}
}
\value{
a list of matrices. Each element corresponds to the cubic spline coefficients for each region.
}
\description{
generateRandomPath
}
