% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestSplitsPath.R
\name{getBestSplitsPath}
\alias{getBestSplitsPath}
\title{Path Model Selection}
\usage{
getBestSplitsPath(EstPath, show, V_Smooth = T, type = "BIC")
}
\arguments{
\item{EstPath}{a list of EstimatedPath Objects. These should all be related using the same data with an increasing number of splits.}

\item{show}{a boolean. TRUE if the criterion values should be displayed.}

\item{V_Smooth}{a boolean. TRUE if the path cubic spline had C2 smoothness. FALSE if C1.}

\item{type}{a string. The model criterion.
Must pick from the following: BIC, BICVel, AIC, VarPos, VarVel, VarPosVel.
BIC and AIC can be used with any path method. BICVel, VarPos, VarVel, VarPosVel can only be used with FullRegression and Projection.}
}
\value{
an EstimatedPath object.
}
\description{
Path Model Selection
}
