% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedPaths_Projection.R
\name{getEstimatedPaths_Projection}
\alias{getEstimatedPaths_Projection}
\title{getEstimatedPaths_Projection}
\usage{
getEstimatedPaths_Projection(
  Data,
  nsplits,
  V_Smooth = T,
  Model_Selection_Type = "VarPosVel"
)
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{nsplits}{a numeric scalar. The maximum number of splits the spline can split into.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}

\item{Model_Selection_Type}{a string. The model criterion.
Must pick from the following: BIC, BICVel, AIC, VarPos, VarVel, VarPosVel.}
}
\value{
an EstimatedPath object.
}
\description{
getEstimatedPaths_Projection
}
\examples{
getEstimatedPaths_Projection(ExData, 100, V_Smooth = FALSE, Model_Selection_Type = "BIC") 

}
