% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedTrajectories_ByFours.R
\name{getEstimatedTrajectories_ByFours}
\alias{getEstimatedTrajectories_ByFours}
\title{getEstimatedTrajectories_ByFours}
\usage{
getEstimatedTrajectories_ByFours(
  EstPath,
  EstVel,
  nsplits,
  baseVectorFields,
  data_int = 0.001,
  Random_Path = F,
  Path_V_Smooth
)
}
\arguments{
\item{EstPath}{an EstimatedPath object.}

\item{EstVel}{an EstimatedPath object. A cubic spline could be fit to estimates of velocity from the data. This is not required.}

\item{nsplits}{a numeric scalar. The maximum number of splits the spline can split into.}

\item{baseVectorFields}{a baseVectorField function.}

\item{data_int}{a numeric scalar. The time grid size used to constrict the trajectory data set.}

\item{Random_Path}{a boolean. TRUE if the path used to construct the trajectory data set should be drawn at random.
If TRUE, the path must've been estimated using the FullRegression or Projection methods.}

\item{Path_V_Smooth}{a boolean. TRUE if the path cubic spline was estimated under C2 smoothness. FALSE if C1. Only needed if Random_Path = TRUE.}
}
\value{
an EstimatedTrajectory object.
}
\description{
By fours method for estimating the trajectory functions.
}
