% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedTrajectories_Optimization.R
\name{getEstimatedTrajectories_Optimization}
\alias{getEstimatedTrajectories_Optimization}
\title{getEstimatedTrajectories_Optimization}
\usage{
getEstimatedTrajectories_Optimization(
  EstPath,
  nsplits,
  baseVectorFields,
  V_Smooth = F,
  Model_Selection_Type = "BICOpt",
  Random_Path = F
)
}
\arguments{
\item{EstPath}{an EstimatedPath object.}

\item{nsplits}{a numeric scalar. The maximum number of splits the spline can split into.}

\item{baseVectorFields}{a baseVectorField object.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}

\item{Model_Selection_Type}{a string. The model criterion.
Must pick from the following: BIC, BICOpt, BICPos, AIC, VarPos.}

\item{Random_Path}{a boolean. TRUE if the path used to estimate the trajectories should be randomly generated.
If TRUE, the path must've been estimated using the FullRegression or Projection method.}
}
\value{
an EstimatedTrajectory object.
}
\description{
getEstimatedTrajectories_Optimization
}
\examples{
ExEstTraj_Optimization = getEstimatedTrajectories_Optimization(
     EstPath = ExEstPath_Projection, 
     nsplits = 100, baseVectorFields = baseVectorFields, 
     V_Smooth = FALSE, Model_Selection_Type = "BICOpt", Random_Path = FALSE)
    
}
