% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathEvaluationFunctions.R
\name{getEstPathAcceleration}
\alias{getEstPathAcceleration}
\title{Evaluate the acceleration of a path}
\usage{
getEstPathAcceleration(t, EstPath)
}
\arguments{
\item{t}{a numeric vector. The times at which the path's acceleration should be evaluated.}

\item{EstPath}{an EstimatedPath object.}
}
\value{
a numeric matrix. Each row corresponds to an evaluated acceleration.
}
\description{
Evaluate the acceleration of a path
}
