% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathEvaluationFunctions.R
\name{getEstPathPosition}
\alias{getEstPathPosition}
\title{Evaluate the position of a path}
\usage{
getEstPathPosition(t, EstPath)
}
\arguments{
\item{t}{a numeric vector. The times at which the path's position should be evaluated.}

\item{EstPath}{an EstimatedPath object.}
}
\value{
a numeric matrix. Each row corresponds to an evaluated position.
}
\description{
Evaluate the position of a path
}
\examples{
getEstPathPosition(c(0,0.5,1), ExEstPath_ByFours)


}
