% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryEvaluationFunctions.R
\name{getEstTrajSlope}
\alias{getEstTrajSlope}
\title{getEstTrajSlope}
\usage{
getEstTrajSlope(t, EstTraj)
}
\arguments{
\item{t}{a numeric vector. The times at which the trajectory's slope should be evaluated.}

\item{EstTraj}{an EstimatedTrajectory object.}
}
\value{
a numeric matrix. Each row corresponds to an evaluated slope.
}
\description{
getEstTrajSlope
}
\examples{
getEstTrajSlope(c(0,0.5,1), ExEstTraj_ByFours)

}
