% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryEvaluationFunctions.R
\name{getEstTrajValue}
\alias{getEstTrajValue}
\title{getEstTrajValue}
\usage{
getEstTrajValue(t, EstTraj)
}
\arguments{
\item{t}{a numeric vector. The times at which the trajectory's value should be evaluated.}

\item{EstTraj}{an EstimatedTrajectory object.}
}
\value{
a numeric matrix. Each row corresponds to an evaluated value.
}
\description{
getEstTrajValue
}
\examples{
getEstTrajValue(c(0,0.5,1), ExEstTraj_ByFours)

}
