% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestSplitsTraj.R
\name{getLikelihoodTraj}
\alias{getLikelihoodTraj}
\title{getLikelihoodTraj}
\usage{
getLikelihoodTraj(EstTraj)
}
\arguments{
\item{EstTraj}{an EstimatedTraj object.}
}
\value{
a scalar.
}
\description{
getLikelihoodTraj
}
