% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedPaths_FullRegression.R
\name{runEstPathStep_FullRegression}
\alias{runEstPathStep_FullRegression}
\title{runEstPathStep_FullRegression}
\usage{
runEstPathStep_FullRegression(Data, time_splits, V_Smooth = T)
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{time_splits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}
}
\value{
an EstimatedPath object.
}
\description{
One iteration in the FullRegression method for estimated paths.
}
