% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedPaths_Projection.R
\name{runEstPathStep_Projection}
\alias{runEstPathStep_Projection}
\title{runEstPathStep_Projection}
\usage{
runEstPathStep_Projection(
  Data,
  time_splits,
  cur_unrestricted_coefs,
  cur_unrestricted_cov,
  cur_integral_list,
  cur_sigma2_list,
  worst_split,
  V_Smooth = T
)
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{time_splits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{cur_unrestricted_coefs}{a list of matrices. Each element corresponds to the coefficient matrices for the unrestricted cubic functions fit on each region of the previous spline.}

\item{cur_unrestricted_cov}{a list of matrices. Each element corresponds to the covariance matrices for the unrestricted cubic coefficients fit on each region of the previous spline.}

\item{cur_integral_list}{a list of matrices. The previous integral list.}

\item{cur_sigma2_list}{a list of vectors. Each element corresponds to the sigma2 estimates of each dimension for the given region of the previous spline.}

\item{worst_split}{a numeric scalar. The region to be split.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}
}
\value{
an EstimatedPath object.
}
\description{
At each iteration an unrestricted cubic function is fit to both halves of the splitting region.
All unrestricted coefficients are projected into the restricted space.
}
