% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedTrajectories_OneShot.R
\name{runEstTrajStep_OneShot}
\alias{runEstTrajStep_OneShot}
\title{runEstTrajStep_OneShot}
\usage{
runEstTrajStep_OneShot(
  EstPath,
  baseVectorFields,
  TimeSplits,
  worst_split,
  cur_integral_list_traj,
  cur_integral_list_pathtraj,
  n_dim,
  n_models,
  Path_VSmooth,
  Traj_VSmooth
)
}
\arguments{
\item{EstPath}{an EtsimatedPath object.}

\item{baseVectorFields}{a baseVectorFields functions.}

\item{TimeSplits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{worst_split}{a numeric scalar. The region to be split.}

\item{cur_integral_list_traj}{a list of matrices. Each corresponds with a block of the current overall trajectory integral matrix.}

\item{cur_integral_list_pathtraj}{a list of matrices. Each corresponds with a block of the current overall PathTrajectory integral matrix.}

\item{n_dim}{a numeric scalar. The number of dimensions in the path.}

\item{n_models}{a numeric scalar. The number of vector fields being mixed.}

\item{Path_VSmooth}{a boolean. TRUE if the path was estimated using C2 smoothness. FALSE if C1.}

\item{Traj_VSmooth}{a boolean. TRUE if the trajectory should be estimated using C2 smoothness. FALSE if C1.}
}
\value{
an EstimatedTrajectory object.
}
\description{
runEstTrajStep_OneShot
}
