% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RungeKutta.R
\name{RungeKutta}
\alias{RungeKutta}
\title{4th-Order Runge-Kutta Method}
\usage{
RungeKutta(
  startTime,
  startPos,
  baseVectorFields,
  TrajList,
  TrajTimeSplits,
  endTime,
  t_step = 0.01
)
}
\arguments{
\item{startTime}{a numeric scalar.}

\item{startPos}{a numeric vector.}

\item{baseVectorFields}{a baseVectorField function.}

\item{TrajList}{a list of numeric matrices. The list of coefficient matrices of the EstimatedTrajectory object.}

\item{TrajTimeSplits}{a numeric vector. The nodes of the EstimatedTrajectory object in addition with the starting and ending time.}

\item{endTime}{a numeric scalar.}

\item{t_step}{a positive numeric scalar. The time grid size for the Runge-Kutta method.}
}
\value{
a numeric matrix. Each row is a time and position of the path through the trajectory-weighted vector field.
}
\description{
4th-Order Runge-Kutta Method
}
\examples{
RungeKutta(startTime = 0, startPos = c(1,1), 
           baseVectorFields = baseVectorFields,
           TrajList = ExTraj$Traj, TrajTimeSplits = ExTraj$TimeSplits, 
           endTime = 5, t_step = 0.01)

}
