% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUnRestrictedCubicRegression_Path.R
\name{runUnRestrictedCubicRegression_Path}
\alias{runUnRestrictedCubicRegression_Path}
\title{runUnRestrictedCubicRegression_Path}
\usage{
runUnRestrictedCubicRegression_Path(cur_Data, start_time, dim)
}
\arguments{
\item{cur_Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{start_time}{a numeric scalar. The centering time for the cubic function being estimated.}

\item{dim}{a numeric scalar. The dimension of the data to estimate the cubic.}
}
\value{
a list. The elements are the estimated coefficients, the estimated value of sigma2,
and the model matrix.
}
\description{
runUnRestrictedCubicRegression_Path
}
