% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEstPaths.R
\name{visualizeEstPath}
\alias{visualizeEstPath}
\title{visualizeEstPath}
\usage{
visualizeEstPath(Data, EstPath, t_grid_size = 0.01, x_var = "t", y_var = "X1")
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{EstPath}{an EstimatedPath object.}

\item{t_grid_size}{a positive numeric scalar. The time grid size of the plotting.}

\item{x_var}{a string. The variable name to plot on the x-axis.}

\item{y_var}{a string. The variable name to plot on the y-axis.}
}
\value{
a ggplot2 object.
}
\description{
visualizeEstPath
}
\examples{
visualizeEstPath(ExData, ExEstPath_ByFours, t_grid_size = 0.01, 
                 x_var = 'X1',y_var = 'X2')
      
}
