% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeFullTrajectories.R
\name{visualizeFullTrajectories}
\alias{visualizeFullTrajectories}
\title{Visualize Trajectory Functions}
\usage{
visualizeFullTrajectories(TrajList, TimeSplitsList, t_grid_size)
}
\arguments{
\item{TrajList}{a list of lists. Each element of the outer list corresponds to a list of cubic matrix coefficients associated with individual cubic functions.}

\item{TimeSplitsList}{a list of numeric vectors. Each element corresponds to nodes of the cubic spline in the current element of TrajList.}

\item{t_grid_size}{a non-negative numeric scalar. The time grid size of the plotting.}
}
\value{
a ggplot2 object.
}
\description{
A function to visualize a series of trajectory functions.
}
\examples{
visualizeFullTrajectories(list(ExTraj$Traj),list(ExTraj$TimeSplits), 0.01)

}
