% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatePathTraj.R
\name{animatePathTraj}
\alias{animatePathTraj}
\title{animatePathTraj}
\usage{
animatePathTraj(
  PathList,
  TrajList,
  PathTimeSplits,
  TrajTimeSplits,
  baseVectorFields,
  n_points
)
}
\arguments{
\item{PathList}{a list of numeric matrices. The list of coefficient matrices of the path.}

\item{TrajList}{a list of numeric matrices. The list of coefficient matrices of the trajectory.}

\item{PathTimeSplits}{a numeric vector. The nodes of the path in addition with the starting and ending time.}

\item{TrajTimeSplits}{a numeric vector. The nodes of the trajectory in addition with the starting and ending time.}

\item{baseVectorFields}{a baseVectorFields function.}

\item{n_points}{the number of points to include in the animation.}
}
\value{
a ggplot2 object. Can be further turned into an animation.
}
\description{
A function to animate a particle traveling through a trajectory-weighted vector field.
}
