% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationFunctions.R
\name{avgMagError}
\alias{avgMagError}
\title{avgMagError}
\usage{
avgMagError(baseVectorFields, Traj, Path, upper, lower)
}
\arguments{
\item{baseVectorFields}{a baseVectorFields function.}

\item{Traj}{an EstimatedTrajectory object.}

\item{Path}{an EstimatedPath object.}

\item{upper}{a numeric scalar. The upper bound for the integral.}

\item{lower}{a numeric scalar. The lower bound for the integral.}
}
\value{
a numeric scalar.
}
\description{
Finds the average magnitude of the error velocity vectors.
}
\examples{
avgMagError(baseVectorFields = baseVectorFields, Traj = ExEstTraj_Optimization, 
            Path = ExEstPath_Projection, upper = 5, lower = 0)
  
}
