% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationFunctions.R
\name{coverageRatePath}
\alias{coverageRatePath}
\title{coverageRatePath}
\usage{
coverageRatePath(EstPath, Data, CL = 0.95)
}
\arguments{
\item{EstPath}{an EstimatedPath object. Must have been estimated using either the FullRegression or Projection methods.}

\item{Data}{a numeric data frame.}

\item{CL}{a real number between 0 and 1. The confidence level of the intervals to check the coverage rate.}
}
\value{
a numeric vector. Each entry corresponds to the coverage rate for a given dimension.
}
\description{
coverageRatePath
}
