% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationFunctions.R
\name{cubicAltPercentError}
\alias{cubicAltPercentError}
\title{cubicAltPercentError}
\usage{
cubicAltPercentError(EstCubic, TrueCubic, EstTimeSplits, TrueTimeSplits)
}
\arguments{
\item{EstCubic}{a list of numeric matrices. The list of coefficient matrices of the estimated cubic spline.}

\item{TrueCubic}{a list of numeric matrices. The list of coefficient matrices of a true cubic spline.}

\item{EstTimeSplits}{a numeric vector. The nodes of the estimated cubic spline in addition with the starting and ending time.}

\item{TrueTimeSplits}{a numeric vector. The nodes of the true cubic spline in addition with the starting and ending time.}
}
\value{
a numeric scalar.
}
\description{
Another cubic percent error-like metric.
}
\examples{
cubicAltPercentError(EstCubic = ExEstTraj_Optimization$Traj, 
                     TrueCubic = ExTraj$Traj, 
                     EstTimeSplits = ExEstTraj_Optimization$TimeSplits, 
                     TrueTimeSplits = ExTraj$TimeSplits)

}
