% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationFunctions.R
\name{cubicPercentError}
\alias{cubicPercentError}
\title{cubicPercentError}
\usage{
cubicPercentError(EstCubic, TrueCubic, EstTimeSplits, TrueTimeSplits)
}
\arguments{
\item{EstCubic}{a list of numeric matrices. The list of coefficient matrices of the estimated cubic spline.}

\item{TrueCubic}{a list of numeric matrices. The list of coefficient matrices of a true cubic spline.}

\item{EstTimeSplits}{a numeric vector. The nodes of the estimated cubic spline in addition with the starting and ending time.}

\item{TrueTimeSplits}{a numeric vector. The nodes of the true cubic spline in addition with the starting and ending time.}
}
\value{
a numeric vector. The first is the percent error and the second is the error associated with the percent error.
}
\description{
Get the percent error between 2 cubic spline functions.
}
