% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drifterVisualizationTool.R
\name{drifterVisualizationTool}
\alias{drifterVisualizationTool}
\title{drifterVisualizationTool}
\usage{
drifterVisualizationTool(
  path,
  traj_list,
  baseVectorFields,
  propagation_list,
  path_t_grid_size,
  traj_t_grid_size,
  model_names,
  method_names,
  model_colors,
  method_colors,
  vector_field_arrow_scale,
  vector_field_grid_resolution
)
}
\arguments{
\item{path}{an EstimatedPath object.}

\item{traj_list}{a list of EstimatedTrajectory objects.}

\item{baseVectorFields}{a baseVectorFields function.}

\item{propagation_list}{a Propagation List.}

\item{path_t_grid_size}{a numeric scalar. The time grid size used to plot the path.}

\item{traj_t_grid_size}{a numeric scalar. The time grid size used to plot the trajectories.}

\item{model_names}{a string vector. The names of the mixing vector fields}

\item{method_names}{a string vector. The names of the methods used to compute the trajectories.}

\item{model_colors}{a string vector. The names or hexcodes of colors to plot the trajectories.}

\item{method_colors}{a string vector. The names or hexcodes of colors to plot the methods.}

\item{vector_field_arrow_scale}{a numeric scalar. The scale used to draw the arrows in Vector Field Mode.}

\item{vector_field_grid_resolution}{a numeric scalar. The number of arrows in either dimension to display the vector fields in Vector Field Mode.}
}
\value{
a shiny app.
}
\description{
A shiny application to take a deep dive into both paths and trajectories.
}
