% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseVectorFieldFunctions.R
\name{flatVF}
\alias{flatVF}
\title{Flat Vector Field Function}
\usage{
flatVF(data_v, baseVectorFields)
}
\arguments{
\item{data_v}{a (d+1)-dimensional numeric vector. The time value is the first entry followed by the position.}

\item{baseVectorFields}{a vector field function.}
}
\value{
a (dxm)-dimensional numeric vector.
The rows of the outputted matrix from baseVectorFields are concatenated.
For example: c(X1v1,X2v1,X1v2,X2v2)
}
\description{
Flattens the matrix output from a vector field function.
}
\examples{
baseVectorFields = function(t, curPos){

f1 = c(curPos[2],-1*curPos[1])
f2 = c(curPos[1],curPos[2])/sqrt(sum(c(curPos[1],curPos[2])^2))

matrix(c(f1,f2), nrow = 2, byrow = FALSE)

}

flatVF(c(0,1,1), baseVectorFields)
}
