% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTrajectoryFunctions.R
\name{generateIndividualTrajectory}
\alias{generateIndividualTrajectory}
\title{Generate Individual Cubic Trajectories}
\usage{
generateIndividualTrajectory(
  startTime,
  endTime,
  prevTraj,
  prevStartTime,
  initTraj = F,
  beta4,
  endSD = 0.5
)
}
\arguments{
\item{startTime}{a numeric scalar. The center time for the generating cubic function.}

\item{endTime}{a numeric scalar. The ending time for the generating cubic function.
If another cubic were to be generated after, this would be the starting time for the next.}

\item{prevTraj}{an nx4 numeric matrix. Each row corresponds to the coefficients of the previous cubic functions.
This argument is not needed if initTraj = TRUE.}

\item{prevStartTime}{a numeric scalar. The centering time for the previous cubic function.
This argument is not needed if initTraj = TRUE.}

\item{initTraj}{a boolean. TRUE if the generating cubic is the first, FALSE if there are cubic functions preceding.}

\item{beta4}{a numeric vector. The constant coefficient for the initial cubic function.
This argument is not needed if initTraj = TRUE.}

\item{endSD}{a non-negative numeric scalar.
The data generating process says the distribution on the difference between the starting and end value of the generated trajectory is normal with mean 0 and standard deviation endSD.}
}
\value{
an nx4 matrix. The generated cubic functions with each row corresponding to the coefficients.
}
\description{
Generate Individual Cubic Trajectories
}
\examples{

startTime <- 2
endTime <- 3
prevTraj <- matrix(c(0,0,0,1), nrow = 1, byrow = TRUE)
prevStartTime <- 1
initTraj <- FALSE
endSD <- 0.5

generateIndividualTrajectory(startTime = startTime, endTime = endTime, 
prevTraj = prevTraj, prevStartTime = prevStartTime, initTraj = initTraj, endSD = endSD)


}
