% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestSplitsTraj.R
\name{getBestSplitsTraj}
\alias{getBestSplitsTraj}
\title{getBestSplitsTraj}
\usage{
getBestSplitsTraj(EstTrajList, EstPath, show = F, type = "BIC", V_Smooth = T)
}
\arguments{
\item{EstTrajList}{a list of EstimatedTraj objects.
They should all be related using the same path with an increasing number of splits.}

\item{EstPath}{an EstimatedPath object. This is the path all the trajectories where estimated from.}

\item{show}{a boolean. TRUE if the criterion values should be displayed.}

\item{type}{a string. The model criterion.
Must pick from the following: AIC, BIC, or BICOpt.
BIC and AIC can be used with the ByFours trajectory method. BICOpt musr be used with Optimization.}

\item{V_Smooth}{a boolean. TRUE if the trajectory cubic splines had C2 smoothness. FALSE if C1.}
}
\value{
an EstimatedTraj object.
}
\description{
getBestSplitsTraj
}
