% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedPaths_ByFours.R
\name{getEstimatedPaths_ByFours}
\alias{getEstimatedPaths_ByFours}
\title{Get Estimated Path using the By Fours Method}
\usage{
getEstimatedPaths_ByFours(Data, nsplits)
}
\arguments{
\item{Data}{a numeric n x (d+1) data frame or matrix. The first column is time with the following each dimension's position.}

\item{nsplits}{a numeric scalar. The maximum number of splits the spline can split into.}
}
\value{
an EstimatedPath object.
}
\description{
One of the functions to estimate a path. By Fours method only fits the previous segment, the two splitting segments, and the next segment during each iteration.
}
