% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedPaths_FullRegression.R
\name{getEstimatedPaths_FullRegression}
\alias{getEstimatedPaths_FullRegression}
\title{getEstimatedPaths_FullRegression}
\usage{
getEstimatedPaths_FullRegression(
  Data,
  nsplits,
  V_Smooth = F,
  Model_Selection_Type = "BIC"
)
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{nsplits}{a numeric scalar. The maximum number of splits the spline can split into.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}

\item{Model_Selection_Type}{a string. The model criterion.
Must pick from the following: BIC, BICVel, AIC, VarPos, VarVel, VarPosVel.}
}
\value{
an EstimatedPath object.
}
\description{
One of the functions to estimate a path. FullRegression method only fits all segments during each iteration.
}
