% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathEvaluationFunctions.R
\name{getEstPathVelocity}
\alias{getEstPathVelocity}
\title{Evaluate the velocity of a path}
\usage{
getEstPathVelocity(t, EstPath)
}
\arguments{
\item{t}{a numeric vector. The times at which the path's velocity should be evaluated.}

\item{EstPath}{an EstimatedPath object.}
}
\value{
a numeric matrix. Each row corresponds to an evaluated velocity.
}
\description{
Evaluate the velocity of a path
}
