% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPropagationList.R
\name{getPropagationList}
\alias{getPropagationList}
\title{getPropagationList}
\usage{
getPropagationList(
  Path,
  TrajList,
  baseVectorFields,
  prop_t_grid_size,
  n_prop_steps
)
}
\arguments{
\item{Path}{an EstimatedPath object.}

\item{TrajList}{a list of EstimatedTrajectory objects.}

\item{baseVectorFields}{a baseVectorFields function.}

\item{prop_t_grid_size}{the time grid size of the propagation points.}

\item{n_prop_steps}{the number of intermediate RungeKutta steps for each propagation.}
}
\value{
a PropagationList
}
\description{
getPropagationList
}
\examples{
getPropagationList(Path = ExEstPath_ByFours, 
             TrajList = list(ExEstTraj_ByFours, ExEstTraj_Optimization), 
             baseVectorFields = baseVectorFields, 
             prop_t_grid_size = 0.1, n_prop_steps = 10)

}
