% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomPathValueFunctions.R
\name{getRandomEstPathPosition}
\alias{getRandomEstPathPosition}
\title{Generate a random position from an Estimated Path}
\usage{
getRandomEstPathPosition(t, EstPath, random = T)
}
\arguments{
\item{t}{a numeric vector. The times at which random positions should be generated.}

\item{EstPath}{an EstimatedPath object. Must have been estimated using the FullRegression or Projection methods.}

\item{random}{a boolean. If TRUE the positions are random.}
}
\value{
a numeric matrix. Each row corresponds to a random position at a time in the t vector.
}
\description{
Generating random positions using the normal distrubutions induced by the linear regression.
}
