% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomPathValueFunctions.R
\name{getRandomEstPathPositionVelocity}
\alias{getRandomEstPathPositionVelocity}
\title{Generate both a random position and velocity from an Estimated Path}
\usage{
getRandomEstPathPositionVelocity(t, EstPath, random = T)
}
\arguments{
\item{t}{a numeric vector. The times at which random positions and velocities should be generated.}

\item{EstPath}{an EstimatedPath object. Must have been estimated using the FullRegression or Projection methods.}

\item{random}{a boolean. If TRUE the positions and velocities are random.}
}
\value{
a numeric matrix. Each row corresponds to a random position and velocity at a time in the t vector.
}
\description{
Generate both a random position and velocity from an Estimated Path
}
\examples{
getRandomEstPathPositionVelocity(c(0,0.5,1), ExEstPath_FullRegression)

}
