% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomPathValueFunctions.R
\name{getRandomEstPathVelocity}
\alias{getRandomEstPathVelocity}
\title{Generate a random velocity from an Estimated Path}
\usage{
getRandomEstPathVelocity(t, EstPath, random = T)
}
\arguments{
\item{t}{a numeric vector. The times at which random velocities should be generated.}

\item{EstPath}{an EstimatedPath object. Must have been estimated using the FullRegression or Projection methods.}

\item{random}{a boolean. If TRUE the velocities are random.}
}
\value{
a numeric matrix. Each row corresponds to a random velocity at a time in the t vector.
}
\description{
Generate a random velocity from an Estimated Path
}
\examples{
getRandomEstPathVelocity(c(0,0.5,1), ExEstPath_FullRegression)

}
