% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnboundedPathRegressionFunctions.R
\name{runUnboundedCubicSplineRegression_Path}
\alias{runUnboundedCubicSplineRegression_Path}
\title{Title}
\usage{
runUnboundedCubicSplineRegression_Path(
  Data,
  dim,
  time_splits,
  split_labels,
  V_Smooth = F
)
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}

\item{dim}{a numeric scalar. The dimension of the data to run the regression.}

\item{time_splits}{a numeric vector. The nodes of the spline with the starting and ending times.}

\item{split_labels}{a numeric or factor vector. Labels for each point in Data for which region of the spline the point falls into.}

\item{V_Smooth}{a boolean. TRUE if the spline should have C2 smoothness. FALSE if C1.}
}
\value{
either a list or NULL. NULL if the model matrix is singular.
If the model matrix is non-singular the list contains the estimates of the coefficients, sigma2, and coefficient covariance matrix.
}
\description{
Title
}
