% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseVectorFieldFunctions.R
\name{TrajWeightedBaseVectorFields}
\alias{TrajWeightedBaseVectorFields}
\title{TrajWeightedBaseVectorFields}
\usage{
TrajWeightedBaseVectorFields(
  t,
  curPos,
  baseVectorFields,
  TrajList,
  TrajTimeSplits
)
}
\arguments{
\item{t}{a numeric scalar. The time the vector fields should be evaluated.}

\item{curPos}{a numeric vector. The position in the vector field.}

\item{baseVectorFields}{a baseVectorFields function.}

\item{TrajList}{a list of numeric matrices. The list of coefficient matrices of the EstimatedTrajectory object.}

\item{TrajTimeSplits}{a numeric vector. The nodes of the EstimatedTrajectory object in addition with the starting and ending time.}
}
\value{
a numeric vector. The velocity vector of the trajectory-weighted vector field at the evaluated point in time and position.
}
\description{
Weight a baseVectorFields function using an EstimatedTrajectory object.
}
