% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFmixing-package.R
\docType{package}
\name{VFmixing-package}
\alias{VFmixing}
\alias{VFmixing-package}
\title{VFmixing: Vector Field Model Mixing}
\description{
Functions to carry out methodology specifically designed to mix vector fields given potentially noisy positional data.
}
\author{
\strong{Maintainer}: Kyle Peterson \email{kdp2abu@virginia.edu}

}
\keyword{internal}
