% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeData.R
\name{visualizeData}
\alias{visualizeData}
\title{visualizeData}
\usage{
visualizeData(Data)
}
\arguments{
\item{Data}{a numeric data frame. The first column is the time and the remaining columns are the positions.}
}
\value{
a ggplot2 object.
}
\description{
visualizeData
}
\examples{
visualizeData(ExData)

}
