% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEstTraj.R
\name{visualizeEstTraj}
\alias{visualizeEstTraj}
\title{visualizeEstTraj}
\usage{
visualizeEstTraj(
  EstTraj,
  TrueTraj,
  TrueTrajTimeSplits,
  EstPath,
  baseVectorFields,
  t_grid_size = 0.01,
  CL = 0.95,
  P_Int = T,
  Cos_Cutoff = 0.91
)
}
\arguments{
\item{EstTraj}{an EstimatedTrajectory object.}

\item{TrueTraj}{a list of numeric matrices. The list of coefficient matrices of a true trajectory. Not required.}

\item{TrueTrajTimeSplits}{a numeric vector. The nodes of the true trajectory in addition with the starting and ending time. Not required.}

\item{EstPath}{an EstimatedPath object. The path used to estimate the trajectories.}

\item{baseVectorFields}{a baseVectorFieldFunction.}

\item{t_grid_size}{a positive numeric scalar. The time grid size of the plotting.}

\item{CL}{a real number between 0 and 1. The confidence level of the intervals to plot. Only used if EstTraj was estimated using the Optimization method.}

\item{P_Int}{a boolean. TRUE if the prediction interval should be plotted. FALSE if the confidence interval on the mean should be plotted.
Only used if EstTraj was estimated using the Optimization method.}

\item{Cos_Cutoff}{a real number between 0 and 1. The cosine value to determine the cutoff for when vector fields are too colinear.}
}
\value{
a ggplot2 object.
}
\description{
visualizeEstTraj
}
